refactor fence code
