+++ packages/core/src/index.ts
@@ -62,3 +62,4 @@ export * from "./parameters"
 export * from "./scripts"
 export * from "./exec"
 export * from "./math"
+export * from "./fence"
