+++ packages/core/src/parsers.ts
@@ -7,7 +7,7 @@ import { MarkdownTrace } from "./trace"
 import { YAMLTryParse } from "./yaml"
 import { DOCXTryParse } from "./docx"
 import { frontmatterTryParse } from "./frontmatter"
-import { extractFenced } from "./template"
+import { extractFenced } from "./fence"
 import { parseAnnotations } from "./annotations"
 import { dotEnvTryParse } from "./dotenv"
 import { INITryParse } from "./ini"
