+++ packages/core/src/template.ts
@@ -199,72 +199,6 @@ export function staticVars(): Omit<ExpansionVariables, "template"> {
     }
 }
 
-export function undoublequote(s: string) {
-    if (s && s[0] === `"` && s[s.length - 1] === `"`) return s.slice(1, -1)
-    return s
-}
-
-const promptFenceStartRx =
-    /^(?<fence>`{3,})(?<language>[^=:]+)?(\s+(?<args>.*))?$/m
-function startFence(text: string) {
-    const m = promptFenceStartRx.exec(text)
-    const groups: Record<string, string> = m?.groups || {}
-    return {
-        fence: groups.fence,
-        language: undoublequote(groups.language),
-        args: parseKeyValuePairs(groups.args),
-    }
-}
-
-export function parseKeyValuePairs(text: string) {
-    const res: Record<string, string> = {}
-    text
-        ?.split(/\s+/g)
-        .map((kv) => kv.split(/[=:]/))
-        .filter((m) => m.length == 2)
-        .forEach((m) => (res[m[0]] = undoublequote(m[1])))
-    return Object.freeze(res)
-}
-
-export function parseVars(vars: string[]) {
-    if (!vars?.length) return undefined
-    const res: Record<string, string> = {}
-    if (vars) for (const v of vars) Object.assign(res, parseKeyValuePairs(v))
-    return Object.freeze(res)
-}
-
-export function renderFencedVariables(vars: Fenced[]) {
-    return vars
-        .map(
-            ({
-                label: k,
-                content: v,
-                validation,
-                args,
-                language,
-            }) => `-   \`${k}\` ${
-                validation !== undefined
-                    ? `schema ${args.schema}: ${validation.valid === undefined ? EMOJI_UNDEFINED : validation.valid ? EMOJI_SUCCESS : EMOJI_FAIL}`
-                    : ""
-            }\n
-\`\`\`\`\`${
-                language ?? /^Note/.test(k)
-                    ? "markdown"
-                    : /^File [^\n]+.\.(\w+)$/m.exec(k)?.[1] || ""
-            }
-${v}
-\`\`\`\`\`
-${
-    validation?.error
-        ? `> [!CAUTION] Schema ${args.schema} validation errors
-${validation.error.split("\n").join("\n> ")}`
-        : ""
-}
-`
-        )
-        .join("\n")
-}
-
 async function parsePromptTemplateCore(
     filename: string,
     content: string,
